+++ src/components/BatchAllocations/BatchAllocationLogs/utils.js
@@ -43,6 +43,7 @@ export const getResultsListFormatter = ({
   intl,
   path,
   locationSearch,
+  locationState,
   disabled,
   selectRecord,
   selectedRecordsMap,
@@ -50,7 +51,7 @@ export const getResultsListFormatter = ({
   return {
     select: (item) => (
       <Checkbox
-        aria-label={intl.formatMessage({ id: 'ui-finance.allocation.batch.logs.columns.select' }, { term: item.id })}
+        aria-label={intl.formatMessage({ id: 'ui-finance.allocation.batch.logs.columns.select' }, { id: item.id })}
         type="checkbox"
         checked={Boolean(selectedRecordsMap[item.id])}
         onChange={() => selectRecord(item)}
@@ -59,7 +60,14 @@ export const getResultsListFormatter = ({
     ),
     [BATCH_ALLOCATION_LOG_FIELDS.jobName]: (item) => (
       <>
-        <TextLink to={`${path}/${item.id}/view${locationSearch}`}>{item.jobName}</TextLink>
+        <TextLink
+          to={{
+            pathname: `${path}/${item.id}/view${locationSearch}`,
+            state: locationState,
+          }}
+        >
+          {item.jobName}
+        </TextLink>
 
         <Button
           onClick={() => exportCsvBatchAllocationLog(item, intl)}
@@ -78,7 +86,7 @@ export const getResultsListFormatter = ({
     [BATCH_ALLOCATION_LOG_FIELDS.createdDate]: (item) => <FolioFormattedTime dateString={item.metadata.createdDate} />,
     [BATCH_ALLOCATION_LOG_FIELDS.updatedDate]: (item) => <FolioFormattedTime dateString={item.metadata.updatedDate} />,
     createdByUsername: (item) => (
-      <TextLink to={`/users/preview/${item.createdByUser.id}`}> {getFullName(item.createdByUser)}</TextLink>
+      <TextLink to={`/users/preview/${item.createdByUser?.id}`}>{getFullName(item.createdByUser)}</TextLink>
     ),
     [BATCH_ALLOCATION_LOG_FIELDS.jobNumber]: (item) => (item.jobNumber || <NoValue />),
   };