+++ src/components/BatchAllocations/hooks/useBatchAllocationLogs/useBatchAllocationLogs.test.js
@@ -37,6 +37,11 @@ const logs = [{
   },
 }];
 
+jest.mock('@folio/stripes-acq-components', () => ({
+  ...jest.requireActual('@folio/stripes-acq-components'),
+  useLocationFilters: jest.fn().mockReturnValue([]),
+}));
+
 const wrapper = ({ children }) => (
   <QueryClientProvider client={queryClient}>
     {children}
@@ -56,7 +61,7 @@ describe('useBatchAllocationLogs', () => {
   });
 
   it('should call logs api endpoint', async () => {
-    const { result } = renderHook(() => useBatchAllocationLogs(), { wrapper });
+    const { result } = renderHook(() => useBatchAllocationLogs({ filters: {} }), { wrapper });
 
     await waitFor(() => expect(result.current.isLoading).toBeFalsy());